requirements.txt
customtkinter>=5.2.0
tkintermapview>=1.0.10
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0
pyogrio>=0.7.0
Pillow>=10.0.0
pytest>=7.0.0
